refactor(manifest): separate `workload.go` declarations (#3218)
This PR refactors the manifest package with the goal of making it easier to figure out in which file to declare a new type.
```
manifest/
├── workload.go # Contains types common across all workloads: rdws, lbws, backend, worker, job
├── workload_ecs.go # Contains types common for ECS workloads: lbws, backend, worker, job
├──  svc.go # Contains common types for services: rdws, lbws, backend, worker
````

Additionally, few minor refactors are present:
- Made private variables that don't need to be exported
- Moved exported variables to functions in order to make them immutable

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.