fix: validate app name before using it to validate svc name in `svc delete` (#3302)
This fixes a bug that surfaces in #3290. We should make sure that `o.appName` is specified (either through flag or prompts) before we use it to validate service name.

This PR also reflects a new coding pattern that we'd like to adopt: validate only optional flags in `Validate` and ask or validate required flag in `Ask`.

In `Ask`, if the prompt is a selection prompt, then after the prompt we shouldn't do existence check. For example, if we need to prompt to select from existing services from application, then we shouldn't validate whether the selected service exists in the application, because this is redundant.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.