chore(pipeline): ensure appName not empty for `pipeline deploy` commands (#3360)
Previously, `tryReadingAppName()` was the default appName value at the cmd.Flags() level.
When we moved that functionality, we introduced a bug:
In the constructors for `SvcListCmd` and `JobListCmd`, we have `appName: vars.appName` and those were empty because we were calling `tryReadingAppName()` later.
This change 1. passes `o.appName` in to those two methods later and 2. gets the appConfig in the same call as validating appName.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.