chore(cli): allow pipeline deletion outside of workspace (#3391)
Fixes https://github.com/aws/copilot-cli/issues/3299.

Previously, we were reading the pipeline manifest 1. to fetch the pipeline name and 2. to look for `access-token-secret`. In order to allow pipeline deletion from outside of the workspace (because it is more flexible and also makes `app delete` more useful), we cannot rely on reading the pipeline manifest because it is inaccessible from outside of the workspace.

Changes in https://github.com/aws/copilot-cli/pull/3382 leveraged the new DeployedPipelines selector in `pipeline delete` so the user can choose the pipeline they would like to delete. This PR solves the remaining problem of how to delete the pipeline secret. Prior to v.1.4 (introduction of CodeStar Connections), Copilot stored GitHub access tokens as secrets in AWS Secrets Manager. Without reading the manifest, we can look for secrets named and tagged with the conventions `pipeline init` followed for GHv1 source connections. If folks changed their pipeline names or secrets tags, we may not catch the secrets. We do log an error saying that we didn't find any secrets, so they are at least aware that no secrets are being deleted. It's obviously better to err on the side of not deleting than deleting. If we do find a secret that we think is a Copilot secret, we prompt for confirmation before proceeding (unless the user passed in the `--delete-secret` flag.

With a pipeline matching the defaults, pipeline name `pipeline-appName-repoName` and secret `github-token-appName-repoName` with tag `Key: "copilot-application"` `Value: [timeStampAtTimeOfInit]`:
```
 % copilot pipeline delete
Found secret 'github-token-muckles-aws-copilot-sample-service'.
Sure? Yes // confirming that the user wants to delete the pipeline

  Are you sure you want to delete the source secret github-token-muckles-aws-copilot-sample-service associated with pipeline pipeline-muckles-aws-copilot-sample-service? [? for help] (y/N) 
```
Without a pipeline match:
```
% copilot pipeline delete
Found no Secrets Manager secrets to delete for pipeline secondPipeline.
Sure? No
✘ pipeline delete cancelled - no changes made
```



By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.