chore(cli): ensure new workloads are not named "pipelines" (#3375)
With the new pipelines file structure, if a new workload is named `pipelines`, it will be potentially confusing with the copilot directory for pipelines. To avoid that issue, this change will ensure new workloads do not use that name.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.