chore(pipelines): make hardcoded paths dynamic in buildspec and pipeline CFN template (#3403)
These changes point the buildspec to the pipeline manifest, whether it's at a legacy location or name-specific location, and also point the pipeline cfn template to the buildspec, whether it's at a legacy or location or a name-specific location. 

An added bonus is that users may indicate, in their pipeline manifests, paths to custom buildspecs.
```
build:
  buildspec: a/path/to/a/different/buildspec.yml
```   

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.