chore: do not trim pipeline stack name to 128 characters (#3421)
This is reverting a part of the changes in #3416 that trim the pipeline stack name to 128 characters by smart-trimming from app name and pipeline name. CFN will error out if the stack name is too long, preventing the initial creation of the stack in the first place. This errorring out is early enough for user to change their pipeline name in order to meet the restriction. 

In addition, the PR updates `pipeline show` and `pipeline status` output.

### `pipeline show`
For `pipeline show`, the human string output will show the pipeline name as specified in manifest:
```
About

  Name        new-world
  Region      us-east-2
  AccountID   1111111111
  Created At  50 minutes ago
  Updated At  50 minutes ago
```
Its JSON output will show:
```
{"pipelineName":"new-world","name":"pipeline-e2e-try-new-world-Pipeline-11T2LFVLO1XN6","region"...
```
adding a new JSON field `pipelineName` whose value is the manifest pipeline name. This is to prevent breaking changes for folks that've been using the JSON output's `name` filed for automation.

### `pipeline status`
`pipeline status` doesn't show pipeline name as human string output, so no changes here.

It's original JSON output is: 
```
{"pipelineName":"pipeline-e2e-try-new-world-Pipeline-11T2LFVLO1XN6"
```
To prevent it from being breaking for users' automation, this PR adds a `name` field whose value is the manifest pipeline name.

Note that this may be confusing because the fields are named differently for `pipeline show` and `pipeline status` JSON.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.