chore(pipeline): allow `pipeline init` to proceed with existing name (#3414)
Instead of erroring out, `pipeline init` will now proceed after logging a warning.
This gives users the flexibility to update their pipeline manifest/buildspec files to the latest defaults by writing the manually deleted files as part of this workflow.

Also in this PR are some changes suggested in the bug bash: using the `--quiet` flag with `wget`, some wording and punctuation changes.

To come: skip prompting for config if reusing existing manifest. Asking for envs, URL, etc. is very misleading if we're not using that info.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.