chore(pipelines): suggest pipeline name to be `[repo]-[branch]` (#3422)
This changes `pipeline init` to suggest `[repo]-[branch]` as a name for the pipeline. Users can press enter to accept the suggested name, or type in a different name if they would like to. This is how the output looks on my machine:

(for branch = `prod`)
![Screen Shot 2022-04-04 at 13 25 15](https://user-images.githubusercontent.com/10566468/161626121-4115fb2f-eee9-45f0-9cc1-721291bf775e.png)

As part of this change, parsing the current branch/repo now occurs before asking for the name, and the output of that shows before asking for the pipeline name, as show in the screenshot above. I think it makes sense in this order, but lmk when you think.

Also refactored the `Execute()` tests to have a single `setupMocks(😇 )` function, and moved tests that dealt with parsing repo/branch to `Ask()`.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.