build: use "go install" install of "get" (#3477)
We can no longer use "go get" to install commands.
Gomega wasn't an executable to begin with so I'm not sure what the "go get" was doing before. 
Removing it and going to see if it fixes the build.


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.