build: publish docs only when a release is published (#3540)
Previously, we were publishing the docs when a "v*" tag was added.
This meant that the docs got out before the github release was public.
This PR changes the trigger action so that only once the release is
actually available publicly, then the docs will be published.


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.