chore: detect if EnvManagerRole has permissions to upload artifacts (#3562)
This PR is step 1 in resolving #3556

We first need to detect if the env manager role is allowed to upload to S3. 
If so, then we're good and can use the env session to upload the template and custom resources.  
Otherwise, in step 2 we'll have to call `UpdateEnvironmentTemplate` by adding the permission to the EnvManagerRole.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.