fix(templates): default path should have lowest priority for HTTPS (#3603)
Currently, for an app with domain, if an LBWS uses the default path that matches everything, all the upcoming LBWS will not get routed for any traffic because of the routing priority. An workaround before this PR would be deploying them in a specific order.


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.