chore: allow non-http svcs in env w/ cert(s) (#3626)
In an environment with imported certs, not all backend services should have to be associated with the internal load balancer. This limits the alias-cert validation to backend services with `http` specified.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.