chore(stack): preserve previous values for env-controller-managed env stack parameters  (#3638)
This PR contains two changes.

Env stack parameters in `Parameters` section can be grouped into those that are controlled by env-controller, and those that are not. The first group of parameter  should only be updated by env controller. This means that the env stack should explicitly assign them `ParameterValue` when these parameters are added to the stack *for the first time*; otherwise, it should `UsePreviousValue`.

An additional change is simplifying part of the changes made by #3528. After v1.14.0, container insights is either "enabled" or "disabled", instead of empty. This behavior should be kept.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.