chore: make windows bin md5 sum file name match others (#3669)
All of the other md5 sums of our output resources are named `filename.md5` except for windows, which has `.exe` stripped off the filename. We don't strip off the `.exe` for the signature files (`.asc`) on windows, so this makes the windows md5 sum match the windows signature file and the rest (linux & mac) of the resources.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.