feat: backend service alb http predictable alias (#3668)
When using HTTP (not HTTPS) for a backend service with an internal ALB, create a hosted zone in form `env.app.internal` and insert an A record `svc.env.app.internal` into the hosted zone pointing to the internal ALB.

Also updates the backend service describer to pull the URI from the listener rule.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.