chore: replace auto-upgrade of environments with compatibility checkers (#3698)
Previously, Copilot automatically upgrades an environment during workload's `deploy` commands if the environment is not on the latest template version.

With this PR, it stops doing that, and instead will check if the workload's manifest is attempting to use a feature that is not yet supported by the environment's template; if yes, then Copilot errors out and asks user to upgrade their environment.

Similar logic applies to the "pipeline + `svc package`" workflow. 

Note: this PR can be merged only after #3688 .

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.