chore(templates): do not use outdated 2008-10-17 IAM policy version
This ensures that our CloudFormation templates use the 2012-10-17 IAM policy version.

Since 2008-10-17 is the default policy version, this also adds the Version key to the policies that were missing it.

Based on https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_policies_elements_version.html, the 2008-10-17 policy should not be used for new policies.

This should not change any behavior, but is likely good practice.