chore(templates): Use consistent form of IAM policy language version
A bare `Version: 2012-10-17` directive is parsed as a Date, but the Version property is expected to be a String. CloudFormation handles this properly, but it's still likely bad form.

This commit makes the policy version field consistently a string - single quotes are used because we don't need escape code behavior.