chore: allow cloudfront distribution creation via env manifest (#3718)
Allows CloudFront to be deployed in front of a load balancer with no certificate/domain set. A following PR will be made to allow the user to configure the environment ALB to allow public traffic to the load balancer or not while CDN is enabled.


PR as part of design in #3701 which satisfies requirements 1, 2, 4, 6, and 12. Milestone 1 still not met with this PR.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.