chore: Document "PipelineBucket" misnomer to avoid confusion (#3748)
I spent a bit too much time reading code & templates not realizing that PipelineBucket was a sort of a misnomer. This change aims to make sure that others reading the code are able to see that it is used for all things that Copilot uploads to S3.



By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.