chore(manifest): add validation and error message for when invalid autoscaling fields set (#3657)
Added a new error type in `errors.go` and create validation for the errors to give a clearer response when invalid scaling fields are set. An example of this is that copilot-cli will now give a descriptive error message when you try to set the `queue_delay` field in the manifest for a `Load Balanced Web Service`.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.