chore(templates): determine http listener rule priority base on http listener (#3769)
The problem with using https listener to determine the priority for `HTTPListenerRuleWithDomain` (a listener rule that's imposed on the HTTP listener, not HTTPS) is, when a service converts from HTTP to using HTTPS, the priority of `HTTPListenerRuleWithDomain` is not necessarily valid.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.