chore(templates): trigger env controller when there is an alias change (#3771)
In addition to the title, also renaming the logical ID of the HTTP rule priority action that is used for `HTTPListenerRuleWithDomain`. This listener rule shouldn't share the same priority action as `HTTPListenerRule`.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.