docs: fix flag description, verb tense, manifest key (#3784)
This PR adds a separate help menu description for `--manifest` for `svc show`, as it wasn't clear that the env name needed to follow the flag for that command.

Also fixed the verb tense for that flag description as well as `--json`; changed affected docs pages. Will try to make the rest consistent at a later date.

Also fixed the `target_container` key name in a sample app on the "Sidecars" page.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.