chore(cfn): Add support for EnvFileARN in task templates (#3762)
This is just a small and straightforward PR that adds support for passing the EnvFileARN in task templates, but does not actually add any user-facing options to add env files. Supporting the actual `--env-file` option will be for a future PR.



Relates to https://github.com/aws/copilot-cli/issues/3672 (part one in a series of PRs)

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.