chore: describe public alb change to allow cloudfront display (#3759)
Allow CloudFront URI to be displayed on top of public ALB. I also changed the struct name from `albURI` to `accessURI` as it better fits the description when the URI is the access point to the  environment, rather than just to the ALB. When PR for prefix lists #3737 is merged, I'll then modify these files to allow to disable showing the ALB DNS altogether.


Addresses #3701 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.