chore: allow cloud front prefix list to restrict security group access via environment manifest (#3737)
Allow users to specify prefix lists, and to restrict ingress traffic to the ALB via the environment manifest. Adds a new field to the PublicHTTPConfig field `restrict_alb_ingress_to_cf` for this behavior.


Feature for part of #3701 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.