refactor(cloudformation): move fileWriter from methods to struct field (#3836)
Previously, each method in the `deploy/cloudformation` pkg took `os.Stderr` as a parameter to render events. 

This refactor moves the parameter to a struct field instead, and makes the progress tracker optional by defaulting to a `discardFile` as the target file for progress events.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.