chore: add cdn certificate to cloudfront and enable feature (#3851)
This adds `cdn.certificate` which allows the user to use certificates imported in `us-east-1` with CloudFront.
There's a few more edge cases needed for the feature to be release ready namely to work with or invalidate `hosted_zone` field in svc manifest with CloudFront.

@iamhopaul123 will be adding the capability for CloudFront to read aliases from the stack parameters to allow a CloudFront distribution creation with imported certs or `--domain` to be able to occur entirely within Copilot


Nearly completes milestone 1 of #3701 and requires #3842 to be merged before merging this

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.