test(regression): add regression test for multi-svc-app (#3853)
This PR sets up regression tests to validate that an application that was functional in an older Copilot version would be still functional in a later version. Right now, there is only one test suite: `"multi-svc-app"` which has a similar but simpler set up as the existing `"multi-svc-app"` test suite in e2e. We will want to add more, for example, regression tests on applications with domain, with imported cert, and with alias.

This PR only contains the test codes, and the make target to dry run the test on local machines. It does not add the test to pipeline. This will be done in a separate PR.


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.