test(integration): retrieve latest tag from a shallow-cloned repo (#3879)
The original way of retrieving the latest tag does not work for a repo that is shallowly-copied (i.e. not full copy). This PR updates the buildspec so that it uses a more robust and cleaner way to do it.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.