test(regression): maintain count of 1 for frontend service (#3893)
Due to the architecture of the application, setting the task count to > 1 introduces flakiness into the e2e test. The job `query` hits the front-end endpoint `/job-setter` that sets an environment variable, and that environment variable is later used to validate that the job `query` indeed has hit the endpoint by checking the response from `/job-checker`. Having count=1 ensures that traffic to both endpoint is routed to the same task.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.