test: fix regression test job failure (#3904)
The `query` job doesn't run ~50% of the time in the regression test suite. This is because we are creating the job with the schedule of `@every 4m`, with the assumption that it will run once on start, and every 4 minutes thereafter. However, after testing, this is not the case - it waits 4 minutes to run the first time.

In the regression test suite, the `query` job is the last one we deploy. Then, we wait a max of 4 minutes to see if the job has run. Since it waits 4 minutes to run after being deployed, it often doesn't get run before we decide it failed. 

This doesn't occur in the e2e test suite because the job is the _second_ (of 4) workloads we deploy. The deployments of the other two services after the job take long enough that the job always has time to execute.

---

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.