fix: preserve existing service discovery endpoint (#3949)
### The bug
Back in v1.9.0, we introduced “multiple environment per VPC”. This feature is shipped with upgrading the service discovery endpoint, from app.local to env.app.local. Since during env upgrade, we preserve all the old parameters, app.local has always been preserved for old environments.

However, env deploy didn’t account for this. It will try to use env.app.local for all environments, old or new. When env deploy is executed against an environment that was created pre-v1.9.0, it results in an attempt to update app.local to env.app.local. This will be breaking for environments that already have services in it, because the service discovery endpoint is exported and imported by the service stack, thus not updatable.

### The fix
The fix simply preserves the value of the  he existing `ServiceDiscoveryEndpoint` parameter. It also accounts for the scenario where `ServiceDiscoveryEndpoint` is not yet in the parameter (pre-v1.9.0 environment that has never been upgraded), and use `app.local` for those environment.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.