fix(cli): use `image.location` value instead of digest on `svc package` (#4055)
Fixes #2912#issuecomment-1265324677

In `svc package`, we previously set a default value of `*string("")` for the `ImageDigest` which caused us to incorrectly think there is a digest available (because we were checking if `digest != nil`).

Fixes by setting the default value of `ImageDigest` to just nil.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.