test: add integ test for permissions boundary (#4048)
This integ test ensures that as we add CFN templates and/or IAM roles to CFN templates, we also conditionally attach a permission boundary if present in the app.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.