chore: add `read_only_fs` as a comment in the default manifest  (#4097)
This PR comments out the `read_only_fs` field that was originally added by default to the manifest. Setting `read_only_fs: true` prevents containers from `RUN`ning command such as `mkdir`.

We still recommend enabling this field. So we add it in the manifest as a comment for visibility.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.