chore: add yaml Union type (#4100)
This PR contains a new generic `Union` type in `package manifest` that can hold one of two underlying types. It also changes the existing type `HealthCheckArgsOrString` to use type `Union[string, HTTPHealthCheckArgs]`. Obviously it will make more sense once we convert _all_ of the yaml union types to use `Union`, but changing it in just this one place should give us chance to test it a bit and make sure we like the implementation.

Some benefits of this change:
- No more custom unmarshal logic for each existing union type
- Can use a single transformer for merging union types

Downsides:
- Struct field access might look a little weird: `type.Field.A` or `type.Field.B.Field`

One bit of weirdness I noticed is that `NewUnionB()` will infer type B, but for some reason the same doesn't happen in `NewUnionA()`. That's why `NewUnionA` has both types but `NewUnionB` has a single type. (unnecessary types are a lint error)

Another callout so it's not missed is that I removed the error when unmarshalling into `HealthCheckArgsOrString` results in an empty `HTTPHealthCheckArgs` struct. This means that previously, we would error with this yaml:
```
http:
  healthcheck:
    random_key: value
```
but now we won't error, the Union value will just be zero or nil (see Union.Unmarshal()). This brings sub fields in line with how we process top level yaml fields - for example, we don't error when unmarshalling this:
```
random_key: value
http:
  path: "/"
```

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.