chore: aurora v2 use correct port for db cluster (#4121)
> The serverless engine mode only applies for Aurora Serverless v1 DB clusters.

https://docs.aws.amazon.com/AmazonRDS/latest/APIReference/API_CreateDBCluster.html

```
Port
The port number on which the DB instances in the DB cluster accept connections.

Default:

When EngineMode is provisioned, 3306 (for both Aurora MySQL and Aurora PostgreSQL)

When EngineMode is serverless:

3306 when Engine is aurora or aurora-mysql

5432 when Engine is aurora-postgresql
```
https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-rds-dbcluster.html#cfn-rds-dbcluster-port



By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.