feat: Add log retention on VPC flow logs (#4164)
This PR resolves #4162  the retention for a log group of VPC FlowLog  i.e, The number of days to retain the log events in the specified log group.

If the flowlogs is `on`
```yaml
network:
  vpc:
    flow_logs: on
```

The default value for retention of log events in log group is set to 14 days[2 weeks].

If the customer want to have a specific retention rate then Environment Manifest can be modified to
```yaml
network:
  vpc:
    flow_logs:
      retention: 30
```
Here are the possible values for retention in days: [https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-logs-loggroup.html](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-logs-loggroup.html)


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.