chore(workspace): fix Use to return an error if the ws has no summary
Currently, running `copilot app init` in an empty workspace results
in the creation of a .workspace file with an empty "" app name.
This is because `Use` did not error on empty workspace and incorrectly
assumed that an empty `copilot/` directory was a valid workspace which
resulted in the app name to be empty.
Instead, we want to prompt for an app name when the workspace is empty.