fix(stream): should not try to fetch from a closed streamer (#4231)
Currently when using progress tracker (especially for tracking StackSet events), it is possible that the program panics because we try to close a closed channel.

Behind the scene, it is because we use
```golang
	waitMs :=
		rand(
			min(
				streamerMaxFetchIntervalDurationMs,
				streamerFetchIntervalDurationMs*(1<<retries),
			)
		)
```
as `Fetch` interval, which is possible to set the wait time to be very close to 0s if `retries` is 0. And then we do the last Fetch (when the stack/stackset/ecs deployment is done), we always set `retries` to 0.

In the caller function `Stream`, we have a select logic as below

```golang
		select {
		case <-ctx.Done():
			return ctx.Err()
		case <-streamer.Done():
			// No more events to Fetch, flush and exit successfully.
			streamer.Notify()
			return nil
		case <-time.After(fetchDelay):
			next, err = streamer.Fetch()
			if err != nil {
				return err
			}
			streamer.Notify()
		}
```
When `fetchDelay` is 0, it could happen that both `streamer.Done()` and `time.After(fetchDelay)` are `true`. And in this scenario, golang will [select one of them randomly](https://go.dev/tour/concurrency/5) (doom when the one with `time.After(fetchDelay)` condition is select).

This fix will set a `streamerMinFetchIntervalDurationMs` to make sure we don't fetch immediately after the previous fetch. It could also help reduce the throttling error we might receive without hurting the current performance.



By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.