docs: ssm secrets do not need to be `SecureString` (#4238)
The current documentation is inaccurate: users do not *have* to store a value as `SecureString` in SSM in order to use it for `secrets` field. As long as it is in SSM, it can be `String` or `SecureString`.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.