refactor: reduce `stack` package's dependency on `deploy` package (#4265)
The refactorsin this PR are:
1. Changing the signature of `CreateAndRenderEnvironment`
2. Move `CreateEnvironmentInput` from `deploy` package to `stack` package
3. Rename `CreateEnvironmentInput` to `EnvConfig`

The original struct `CreateEnvironmentInput` is only used to construct `stack.Env`. Therefore, it shouldn't live in `deploy` package, and should instead be considered an input struct to `stack.Env`'s constructor.

This is a peripheral effort to support #4219.


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.