refactor: manifest pkg to support multiple ports (#4268)
This PR refactors the manifest package in order to create a list of exposed ports. Though we are still accepting a single port from the customer. Also introducing new field `target_port` under `http` that can open another port internally from either primary container or sidecar container. In this PR because there are no CFN changes so nothing really reflects in the CFN. 

This PR is a part of #4192 


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.