chore: reorder addons storage code (#4330)
This PR contains 0️⃣  actual code changes. It is merely just moving codes around so that struct/func definition are grouped in a better way. This PR is made so that a follow-up PR does not contain these irrelevant line changes.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.