chore(manifest): make cdn static path to be required (#4331)
Making cdn static path to be required because ALB is the default cache behavior and if we give `*` as a default value for path, it means all the traffic will be routed to the static pattern (a possible use case but not a sane default).


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.